Update to typescript
