--- src/theme/DocSidebar/Desktop/Content/index.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="react" />
-import type { Props } from '@theme/DocSidebar/Desktop/Content';
-export default function DocSidebarDesktopContent({ path, sidebar, className, }: Props): JSX.Element;
